requirements.txt
streamlit>=1.31.0
pandas>=2.1.4
plotly>=5.18.0
duckdb>=0.8.1
boto3>=1.34.14
python-dotenv>=1.0.0
numpy>=1.26.3
scikit-learn>=1.3.2
matplotlib>=3.8.2
seaborn>=0.13.1
pyarrow>=14.0.2
openpyxl>=3.1.2
